+++ CHANGELOG.md
@@ -2,6 +2,47 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.3.0-rc.9"></a>
+# [0.3.0-rc.9](https://github.com/SAP/fundamental-react/compare/v0.3.0-rc.8...v0.3.0-rc.9) (2019-01-31)
+
+
+### Bug Fixes
+
+* DatePicker update unit tests ([#332](https://github.com/SAP/fundamental-react/issues/332)) ([7fffa2a](https://github.com/SAP/fundamental-react/commit/7fffa2a))
+* issue [#181](https://github.com/SAP/fundamental-react/issues/181) DatePicker disabled date selected ([#327](https://github.com/SAP/fundamental-react/issues/327)) ([68899f0](https://github.com/SAP/fundamental-react/commit/68899f0))
+
+
+
+<a name="0.3.0-rc.8"></a>
+# [0.3.0-rc.8](https://github.com/SAP/fundamental-react/compare/v0.3.0-rc.7...v0.3.0-rc.8) (2019-01-31)
+
+
+
+<a name="0.3.0-rc.7"></a>
+# [0.3.0-rc.7](https://github.com/SAP/fundamental-react/compare/v0.3.0-rc.6...v0.3.0-rc.7) (2019-01-31)
+
+
+
+<a name="0.3.0-rc.6"></a>
+# [0.3.0-rc.6](https://github.com/SAP/fundamental-react/compare/v0.3.0-rc.5...v0.3.0-rc.6) (2019-01-31)
+
+
+### Bug Fixes
+
+* lint scripts ([#328](https://github.com/SAP/fundamental-react/issues/328)) ([bef19a0](https://github.com/SAP/fundamental-react/commit/bef19a0))
+
+
+
+<a name="0.3.0-rc.5"></a>
+# [0.3.0-rc.5](https://github.com/SAP/fundamental-react/compare/v0.3.0-rc.4...v0.3.0-rc.5) (2019-01-31)
+
+
+### Bug Fixes
+
+* Calendar component tests failing in Travis ([#329](https://github.com/SAP/fundamental-react/issues/329)) ([6817d31](https://github.com/SAP/fundamental-react/commit/6817d31))
+
+
+
 <a name="0.3.0-rc.4"></a>
 # [0.3.0-rc.4](https://github.com/SAP/fundamental-react/compare/v0.3.0-rc.3...v0.3.0-rc.4) (2019-01-30)